refactor(cli): split onlyPlugins functions, improve unit and integration tests
